fix(laravel): address PR review feedback for config-based default spec
- Validate config() return type with is_string guard for type safety
- Improve error message to mention config key and override option
- Update CLAUDE.md Laravel Integration section for new architecture
- Improve LaravelConfigMock docblock (correct package attribution,
  explain namespace mocking rationale, warn about use function imports)
- Extract shared CreatesTestResponse trait from duplicated helpers
- Add null config value test and update error message assertions
- Add explicit config() ignore in phpstan.neon.dist for trait file
