+++ CLAUDE.md
@@ -49,7 +49,8 @@ This is a PHP library (`studio-design/openapi-contract-testing`) that validates 
 
 ### Laravel Integration
 
-- **`ValidatesOpenApiSchema`** trait — Used in Laravel test cases. Provides `assertResponseMatchesOpenApiSchema()` which auto-resolves method/path from the current request and records coverage. Requires implementing `openApiSpec(): string` to specify which spec to validate against.
+- **`ValidatesOpenApiSchema`** trait — Used in Laravel test cases. Provides `assertResponseMatchesOpenApiSchema()` which auto-resolves method/path from the current request and records coverage. The default spec name is read from `config('openapi-contract-testing.default_spec')`; override `openApiSpec(): string` per-test-class if needed.
+- **`OpenApiContractTestingServiceProvider`** — Auto-discovered service provider that registers and publishes the `openapi-contract-testing` config file (`default_spec` key).
 
 ### Key Enums
