+++ phpstan.neon.dist
@@ -6,7 +6,7 @@ parameters:
     tmpDir: .phpstan.cache
     ignoreErrors:
         -
-            message: '#Function app not found#'
+            message: '#Function (app|config) not found#'
             path: src/Laravel/ValidatesOpenApiSchema.php
         -
             message: '#Call to an undefined method Illuminate\\Testing\\TestResponse::#'
